refactor: unify api service
