+++ frontend/src/hooks/useAPIStatus.ts
@@ -1,12 +1,12 @@
 import { useEffect, useState } from "react";
-import { api } from "../lib/api";
+import apiService from "@/services/apiService";
 
 export function useAPIStatus() {
   const [status, setStatus] = useState<string | null>(null);
   const [error, setError] = useState<string | null>(null);
 
   useEffect(() => {
-    api
+    apiService
       .get("/")
       .then((res) => setStatus(res.data.message))
       .catch((err) => setError(err.message));
